[mips] Fix snapshot generation on qemu-user
When making a big-endian MIPS snapshot on a little-endian system, the
address space in a qemu-user process is typically so fragmented that
a 2 GiB reservation is too much to ask. By lowering the reservation
size to 1 GiB we avoid this issue, and still have plenty of headroom
for large snapshots.